Restructure the QueryBoundary directory
This is easier to navigate as it groups together all the context-related components together.
