+++ src/deprecated/LoaderData.tsx
@@ -1,6 +1,6 @@
-import type { QueryBoundaryDataProps } from "src/providers";
+import type { QueryBoundaryDataProps } from "src/groups";
 
-import { QueryBoundaryData } from "src/providers";
+import { QueryBoundaryData } from "src/groups";
 
 /** @deprecated This type has been renamed to QueryBoundaryDataProps. */
 export type LoaderDataProps<DataType> = QueryBoundaryDataProps<DataType>;
