+++ src/groups/QueryBoundary/QueryBoundaryError.tsx
@@ -3,7 +3,7 @@ import type { ReactNode } from "react";
 import Alert from "@mui/material/Alert";
 import { useRef } from "react";
 
-import { useQueryBoundary } from "src/providers/QueryBoundaryProvider/QueryBoundaryProvider";
+import { useQueryBoundary } from "src/groups/QueryBoundary/QueryBoundaryProvider";
 
 export interface QueryBoundaryErrorProps {
   /** The component to show if an error has been thrown. */
