+++ src/groups/QueryBoundary/QueryBoundaryFallback.tsx
@@ -1,10 +1,10 @@
 import type { ReactNode } from "react";
 
-import type { QueryBoundaryErrorProps } from "src/providers/QueryBoundaryProvider/QueryBoundaryError";
-import type { QueryBoundaryNullableProps } from "src/providers/QueryBoundaryProvider/QueryBoundaryNullable";
+import type { QueryBoundaryErrorProps } from "src/groups/QueryBoundary/QueryBoundaryError";
+import type { QueryBoundaryNullableProps } from "src/groups/QueryBoundary/QueryBoundaryNullable";
 
-import QueryBoundaryError from "src/providers/QueryBoundaryProvider/QueryBoundaryError";
-import QueryBoundaryNullable from "src/providers/QueryBoundaryProvider/QueryBoundaryNullable";
+import QueryBoundaryError from "src/groups/QueryBoundary/QueryBoundaryError";
+import QueryBoundaryNullable from "src/groups/QueryBoundary/QueryBoundaryNullable";
 
 export type QueryBoundaryFallbackProps = Omit<QueryBoundaryErrorProps, "children"> & {
   /** The component to show if an error has been thrown. */
