+++ src/hooks/index.ts
@@ -1,12 +1,2 @@
-export { default as createQueryBoundary } from "src/hooks/createQueryBoundary";
 export { default as useDebounce } from "src/hooks/useDebounce";
 export { default as useHash } from "src/hooks/useHash";
-
-export type {
-  Query,
-  QueryBase,
-  QueryMultiple,
-  QuerySingle,
-  CreateQueryBoundaryParameters,
-  DefaultQueryBoundaryComponents,
-} from "src/hooks/createQueryBoundary";
