+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from "src/components";
 export * from "src/deprecated";
+export * from "src/groups";
 export * from "src/hooks";
 export * from "src/providers";
