--- src/providers/QueryBoundaryProvider/index.tsx
@@ -1,31 +1,0 @@
-import QueryBoundaryProvider from "src/providers/QueryBoundaryProvider/QueryBoundaryProvider";
-
-export { default as QueryBoundaryData } from "src/providers/QueryBoundaryProvider/QueryBoundaryData";
-export { default as QueryBoundaryDataMap } from "src/providers/QueryBoundaryProvider/QueryBoundaryDataMap";
-export { default as QueryBoundaryError } from "src/providers/QueryBoundaryProvider/QueryBoundaryError";
-export { default as QueryBoundaryFallback } from "src/providers/QueryBoundaryProvider/QueryBoundaryFallback";
-export { default as QueryBoundaryNullable } from "src/providers/QueryBoundaryProvider/QueryBoundaryNullable";
-export { useQueryBoundary } from "src/providers/QueryBoundaryProvider/QueryBoundaryProvider";
-
-export type { QueryBoundaryDataProps } from "src/providers/QueryBoundaryProvider/QueryBoundaryData";
-export type {
-  QueryBoundaryDataMapBaseProps,
-  QueryBoundaryDataMapProps,
-  QueryBoundaryDataMapPropsWithDataParser,
-  QueryBoundaryDataMapPropsWithItemParser,
-  QueryBoundaryDataMapPropsWithNoParser,
-} from "src/providers/QueryBoundaryProvider/QueryBoundaryDataMap";
-export default QueryBoundaryProvider;
-export type { QueryBoundaryErrorProps } from "src/providers/QueryBoundaryProvider/QueryBoundaryError";
-export type { QueryBoundaryFallbackProps } from "src/providers/QueryBoundaryProvider/QueryBoundaryFallback";
-export type {
-  QueryBoundaryNullablePropsWithNullable,
-  QueryBoundaryNullablePropsWithUndefinedOrNull,
-  QueryBoundaryNullableProps,
-} from "src/providers/QueryBoundaryProvider/QueryBoundaryNullable";
-export type {
-  QueryBoundaryProviderProps,
-  QueryBoundaryProviderBaseProps,
-  QueryBoundaryProviderPropsWithError,
-  QueryBoundaryProviderPropsWithNoError,
-} from "src/providers/QueryBoundaryProvider/QueryBoundaryProvider";
