+++ src/providers/index.ts
@@ -1,9 +1,6 @@
 export { default as ModeProvider, useMode } from "src/providers/ModeProvider";
-export { default as QueryBoundaryProvider } from "src/providers/QueryBoundaryProvider";
 export { default as ScreenSizeProvider, useScreenSize } from "src/providers/ScreenSizeProvider";
 export { default as SnackbarProvider, useSnackbar } from "src/providers/SnackbarProvider";
-
-export * from "src/providers/QueryBoundaryProvider";
 
 export type { ModeProviderProps } from "src/providers/ModeProvider";
 export type { ScreenSizeProps, ScreenSizeContextValue } from "src/providers/ScreenSizeProvider";
